slide Content component for separate re rendering, uploading fixing, can get userId from token after login, dynamic render to avoid hydration error
