+++ app/uploading/page.tsx
@@ -4,8 +4,10 @@ import { Upload } from "lucide-react";
 import React, { useState } from "react";
 
 const Page = () => {
-  const serverUrl = process.env.NEXT_PUBLIC_SERVER_URL || "http://localhost:4000";
+  const serverUrl =
+    process.env.NEXT_PUBLIC_SERVER_URL || "http://localhost:4000";
   const [file, setFile] = useState<File | null>(null);
+  const [previewUrl, setPreviewUrl] = useState<string | null>(null);
 
   const upload = async () => {
     try {
@@ -34,9 +36,16 @@ const Page = () => {
           const files = e.target.files;
           if (files && files.length > 0) {
             setFile(files[0]);
+            const imageUrl = URL.createObjectURL(files[0]);
+            setPreviewUrl(imageUrl);
           }
         }}
       />
+      {previewUrl && (
+        <div>
+          <img src={previewUrl} className="max-w-xs rounded mt-4"/>
+        </div>
+      )}
       <button
         className="border-2 px-4 py-2 flex items-center space-x-2 bg-blue-500 text-white hover:bg-blue-600"
         type="button"
